requirements.txt
torch
torchvision
transformers
lpips
Pillow
requests
numpy
scikit-image
faiss-cpu
streamlit